+++ apps/factory/ui/settings/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useTheme'
 export * from './useVscodeApi'
 export * from './usePanelVisibility'
+export * from './useNow'